restructures naming for routes
